requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.11.29
djangorestframework==3.9.4
idna==2.10
importlib-metadata==2.1.3
kombu==4.6.11
pathlib2==2.3.7.post1
pytz==2022.6
requests==2.27.1
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
urllib3==1.26.12
vine==1.3.0
zipp==1.2.0
